Screen reader view
New generation option "screen reader". Generates map as an HTML table. Currently, details panel on tables may not be the best in terms of accessibility, but for now it's what works. I will figure out if the solution to render details panel mid table is ok, otherwise I will work out a fix.